% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_source_decomp_comp.R
\name{gini.source.decomp.comp}
\alias{gini.source.decomp.comp}
\title{Gini decomposition components}
\usage{
gini.source.decomp.comp(.data, ..., .by = NULL, .wgt = NULL)
}
\arguments{
\item{.data}{A data frame, or data frame extension (e.g. a tibble)}

\item{...}{One or more unquoted expressions separated by commas indicating income sources to consider in the decomposition. Variable names can be used as if they were positions in the data frame.}

\item{.by}{A column to group the calculations by.}

\item{.wgt}{an optional vector of weights to apply in computation. Should be NULL or a numeric vector.}
}
\value{
A tibble containing three components (for each source) of income decomposition.
}
\description{
This function calculates the Gini correlation between two distributions.
}
