% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_correlation.R
\name{gini_corr}
\alias{gini_corr}
\title{Gini correlation}
\usage{
gini_corr(x, y, weights = NULL)
}
\arguments{
\item{x}{a numeric vector containing at least non-negative elements.}

\item{y}{a numeric vector containing the distribution with the rank information.}

\item{weights}{an optional vector of weights to apply in computation. Should be NULL or a numeric vector.}
}
\value{
The value of the Gini correlation, should be between -1 and 1.
}
\description{
This function calculates the Gini correlation between two distributions.
}
