% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_decomposition_funs.R
\name{gini_income_elasticity}
\alias{gini_income_elasticity}
\title{Gini income elasticity}
\usage{
gini_income_elasticity(.data, ..., .by = NULL, .wgt = NULL)
}
\arguments{
\item{.data}{A data frame, or data frame extension (e.g. a tibble)}

\item{...}{One or more unquoted expressions separated by commas indicating income sources to consider in the decomposition. Variable names can be used as if they were positions in the data frame.}

\item{.by}{A column to group the calculations by.}

\item{.wgt}{an optional vector of weights to apply in computation. Should be NULL or a numeric vector.}
}
\value{
A tibble containing the value of elasticity and the marginal effect of each source on income inequality.
}
\description{
This function computes the elasticity of Gini index associated with a percentage change in the mean income (for each income source).
}
