% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_wtd_ord.R
\name{gini_wtd_ord}
\alias{gini_wtd_ord}
\title{Gini index calculation (option of enforcing order)}
\usage{
gini_wtd_ord(x, ord = NULL, weights = NULL)
}
\arguments{
\item{x}{a numeric vector containing at least non-negative elements.}

\item{ord}{a numeric vector containing the distribution to sort with.}

\item{weights}{an optional vector of weights of x to be used in the computation of the Gini coefficient. Should be NULL or a numeric vector.}
}
\value{
The value of the Gini index.
}
\description{
This function calculates the Gini index for a distribution, with an option of sorting according to another distribution.
}
