% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_income_data}
\alias{sample_income_data}
\title{Sample income data}
\format{
A data frame with 200 rows and 6 columns:
\describe{
\item{region}{Whether the households residence is urban or rural}
\item{sample_wgt}{Sample weight}
\item{wage}{Wage and salary}
\item{self_employment_rev}{Earnings from self-employment }
\item{farming_rev}{Income derived from agricultural activities}
\item{other_rev}{Other income sources}
...
}
}
\usage{
sample_income_data
}
\description{
A simulated data for households income sources
}
\keyword{datasets}
