% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_decomposition_funs.R
\name{gini_decomp_source}
\alias{gini_decomp_source}
\title{Gini decomposition by income sources}
\usage{
gini_decomp_source(.data, ..., .by = NULL, .wgt = NULL)
}
\arguments{
\item{.data}{A data frame, or data frame extension (e.g. a tibble)}

\item{...}{One or more unquoted expressions separated by commas indicating income sources to consider in the decomposition. Variable names can be used as if they were positions in the data frame.}

\item{.by}{A column to group the calculations by.}

\item{.wgt}{an optional vector of weights to apply in computation. Should be NULL or a numeric vector.}
}
\value{
A tibble containing contribution of each source to the income inequality.
}
\description{
This function provides a decomposition of Gini index by income sources based on the approach of Lerman and Yitzhaki (1985).
}
